upgrading FileLoader tests for structure and readability
Utilizing the describe element for structured tiers of test steps,
then tests for the line-item checks.  This only provides one test
so far, for the rendering of the React Fragment, and verifies two
properties separately.

In this style of testing, with line-item checks, it is easier to see
the minor steps that are possibly contributing to bugs, or isolate
things that still work as intended.