setting up testing for WordLoader
building off of the changes already pending for the FileLoader, I set up
testing for WordLoader to be able to start designing thoughtful tests.

First, one of the warnings / errors with jest was that there was a rule
for mocking css styles, but no file at the location to handle them.
I added styleMock.js to address that.

Next, the index file was not exporting the WordLoader class specifically
so I added the export keyword to do so (and be able to import it for
testing)

Then I mostly duplicated the FileLoader test file for WordLoader, but
rewrote the render check to have some appropriate checks for this class